% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{asymmetric.mad}
\alias{asymmetric.mad}
\title{Asymmetric NMAD thresholds}
\usage{
asymmetric.mad(dat, n = 6)
}
\arguments{
\item{dat}{Data for which bounds are to be calculated}

\item{n}{Number of multiples of MAD to be applied either side of mu}
}
\value{
Vector containing upper and lower bounds
}
\description{
Calculate modal density mu of data and find boundaries at distance n * MAD either side of mu.
}

