% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_All-the-functions.R
\name{bad.px.limits}
\alias{bad.px.limits}
\title{For a single image array, return a list of thresholds to categorise very bright, bright, slightly bright, slightly dim, dim, and very dim pixels. Edge pixels are ignored for the purpose of this calculation.
Get cutoffs for bright/dim pixels}
\usage{
bad.px.limits(im, inner = "Johnson", edge.width = 10)
}
\arguments{
\item{im}{Two-dimensional image array containing pixel values to be thresholded}
}
\value{
List of named numbers to use as cutpoints
}
\description{
For a single image array, return a list of thresholds to categorise very bright, bright, slightly bright, slightly dim, dim, and very dim pixels. Edge pixels are ignored for the purpose of this calculation.
}
\examples{
img <- pw.m[,,"white", "160314"]
table(findInterval(img, unlist(bad.px.limits(img))))

}

