% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thresholding2.R
\name{bad.px}
\alias{bad.px}
\title{Get bad pixel map for a given date}
\usage{
bad.px(dt)
}
\arguments{
\item{dt}{Integer or string giving acquisition date (format yymmdd)}
}
\description{
Use thresholding and absolute values to classify bad pixels in a single acquisition, using the thresholding limits set in \link{\code{bad.bx.limits}}.
}
\details{
Requires that the pixelwise mean array \code{pw.m} and a residual array \code{res} be available in the global environment.
}
\examples{
bp <- bad.pixels("160314")

}

