% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary-operations.R
\name{batch.summary}
\alias{batch.summary}
\title{Get summary statistics for an image batch}
\usage{
batch.summary(data)
}
\arguments{
\item{data}{Array with 3 or more dimensions}
}
\value{
Vector containing named summary statistics
}
\description{
Return a vector of summary statistics for a batch of acquisitions: mean, standard deviation, median, upper and lower quartiles, interquartile range, and max and min pixelwise mean value.
}
\details{
Pixelwise max and min value are used to check if image is scaled from absolute 0:1 or from an arbitrary minimum point
}
\examples{
b.150828.summ <- batch.summary(b.150828)


}

