% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thresholding.R
\name{bp.by.res.sd}
\alias{bp.by.res.sd}
\title{Identify extreme values along transect}
\usage{
bp.by.res.sd(transect, loess.span = 1/15, keep = 0.95, display = T,
  legend = F, legend.pos = "topright", ...)
}
\arguments{
\item{transect}{Vector of pixel values to be assessed for 'bad' pixels}

\item{loess.span}{Filter span to be used in smoothing the data. Default is 1/15.}

\item{keep}{Central proportion of points to keep. Default is 0.95, which results in a cutoff for 'bad' pixels of 1.96sd.}

\item{display}{Boolean: plot the resulting transect? Default is T.}

\item{legend}{Boolean: include legend in plot? Default is F.}

\item{legend.pos}{String: where to display legend? default is "topright".}

\item{...}{Additional parameters to be passed to plotting function}
}
\value{
A list containing named vectors of Loess-smoothed values, residuals, and indices of pixels flagged.
}
\description{
Find extreme pixel values in single vector
}
\details{
Uses Loess smoothing to find 'spine' of vector and identifies any points that fall more than a certain constant multiple of the SD from the central spine.
}
\examples{
zz <- bp.by.res.sd(pw.m[992, 1:992], display = T)
}

