% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{bp.colours}
\alias{bp.colours}
\title{Colour scheme for bad pixel plot}
\usage{
bp.colours(block = c("edge", "s.bright"))
}
\arguments{
\item{block}{Vector of categories to omit from the plot, which will be set to colour NA. Default is c("edge", "s.bright").}
}
\value{
Vector of colours.
}
\description{
Support function: returns a vector of colours to be used in bad pixel plotting.
}

