% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_All-the-functions.R
\name{bp.lowess.johnson}
\alias{bp.lowess.johnson}
\title{Identify 'bad' pixels using quantiles of Lowess residuals}
\usage{
bp.lowess.johnson(img.date, batch, span = 1/15, lq = 0.001, uq = 0.999,
  method = "quant", details = F)
}
\arguments{
\item{img.date}{Integer/string specifying image date to be imported}

\item{batch}{String specifying colour batch to be imported (black, grey or white)}

\item{span}{Smoothing span to be used for Lowess. Default is 1/15}

\item{lq}{Lower quartile at which to start looking for 'outliers'. Default is 0.001}

\item{uq}{Upper quartile at which to start looking for 'outliers'. Default is 0.999}

\item{method}{String: fit Johnson distribution by quantiles or by matching moments? Default is "quant"}

\item{details}{Boolean: return lists of identified 'bad pixels' (T) or only summary (F)? Default is F.}
}
\description{
Load a specified image and apply Lowess smoothing along columns, then fit a Johnson model and use its quantiles to extract residuals
}
\details{
Requires that pixelwise mean files are already loaded, as by \link{\code{load.pixel.maps}}
}
\examples{
load.pixel.maps()
zz <- rbind.fill(bp.lowess.quantiles(160314, "white"),
                 bp.lowess.johnson(160314, "white"))
                 
}

