% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_All-the-functions.R
\name{bp.lowess.quantiles}
\alias{bp.lowess.quantiles}
\title{Identify 'bad' pixels using quantiles of Lowess residuals}
\usage{
bp.lowess.quantiles(img.date, batch, span = 1/15, lq = 0.001, uq = 0.999,
  dist = 1.5, details = F)
}
\arguments{
\item{img.date}{Integer/string specifying image date to be imported}

\item{batch}{String specifying colour batch to be imported (black, grey or white)}

\item{span}{Smoothing span to be used for Lowess. Default is 1/15}

\item{lq}{Lower quartile at which to start looking for 'outliers'. Default is 0.001}

\item{uq}{Upper quartile at which to start looking for 'outliers'. Default is 0.999}

\item{dist}{Multiples of IQR away from lq/uq at which a point is classified as 'bad'.}

\item{details}{Boolean: return lists of identified 'bad pixels' (T) or only summary (F)? Default is F.}
}
\description{
Load a specified image and apply Lowess smoothing along columns, then cut residuals at specified quantiles to classify 'bad' pixels
}
\details{
Requires that pixelwise mean files are already loaded, as by \link{\code{load.pixel.maps}}
}
\examples{
load.pixel.maps()
zz <- rbind(bp.lowess.quantiles(160314, "white"),
            bp.lowess.quantiles(151015, "white"),
            bp.lowess.quantiles(150828, "white"))

}

