% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{bpx2im}
\alias{bpx2im}
\title{Convert bad pixel map to image}
\usage{
bpx2im(bpx, im.dim = c(1996, 1996))
}
\arguments{
\item{bpx}{Bad pixel map: data frame with first two columns containing x and y coordinates of bad pixels, and a column 'type' specifying what kind of bad pixel was identified.}

\item{im.dim}{Vector of dimensions of image array to create. Default is c(1996, 1996).}
}
\value{
2d image array of the specified dimensions
}
\description{
Given a bad pixel map with coordinates and types, create a 2d array showing each type of bad pixel identified
}

