% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thresholding.R
\name{bvn}
\alias{bvn}
\title{Calculate bivariate normal density of image surface}
\usage{
bvn(par, obs)
}
\arguments{
\item{par}{Vector of named parameters A (constant offset), x0 (x-origin), y0 (y-origin), sig.x (SD over x), sig.y (SD over y)}

\item{obs}{Data frame with columns x and y containing coordinates at which to evaluate the function.}
}
\value{
Vector containing calculated density
}
\description{
Calculate the bivariate Gaussian density of an image surface with given parameters (rho fixed at 0)
}
\examples{
gv <- setNames(melt(pw.m[,,"grey", "160430"]), nm = c("x", "y", "z"))
tmp <- bvn(c(A = 15000, x0 = 1024.5, y0 = 1024.5, sig.x = 500, sig.y = 500), gv)
pixel.image(array(tmp, dim = c(2048, 2048)))

}

