% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thresholding.R
\name{bvn.ss}
\alias{bvn.ss}
\title{Support function: get sum of squared errors of bivariate normal density}
\usage{
bvn.ss(par, obs)
}
\arguments{
\item{par}{Vector of named parameters, to be passed to function to be evaluated}

\item{obs}{Data frame containing columns named x, y and z, containing coordinates at which function should be evaluated, and observed values at those points.}

\item{fn}{Function to be evaluated}
}
\description{
Function to calculate bivariate Gaussian density (with rho = 0) with given parameters for image surface, and return sum of squared residuals. Primarily a support function for \link{\code{optim}}.
}

