% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thresholding.R
\name{class.boundaries}
\alias{class.boundaries}
\title{Calculate class boundaries}
\usage{
class.boundaries(dat, b = c(0.25, 0.5))
}
\arguments{
\item{dat}{Data for which boundaries are to be calculated}

\item{b}{Boundaries to calculate (as well as \code{\link{asymmetric.mad}}): default is c(0.25, 0.5)}
}
\value{
Vector of cutpoints
}
\description{
For a pixelwise mean image, calculate thresholds for each classification
}

