% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thresholding.R
\name{classify.px}
\alias{classify.px}
\title{Label extreme-valued pixels}
\usage{
classify.px(im, levels = c("dim", NA, "s.bright", "bright", "v.bright"))
}
\description{
Full explanation still required
}

