% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_All-the-functions.R
\name{classify.states}
\alias{classify.states}
\title{Classify bad pixels according to the objects to which they belong.}
\usage{
classify.states(bpm)
}
\arguments{
\item{bpm}{Bad pixel map}
}
\value{
Updated bad pixel map
}
\description{
Given a bad pixel map, classify singletons, clusters and superclusters according to their composition.
}

