% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{cluster.px.old}
\alias{cluster.px.old}
\title{Identify clusters and superclusters of pixels (old version of function, held for reference only)}
\usage{
cluster.px.old(bpx)
}
\arguments{
\item{bpx}{Bad pixel map: a matrix or data frame containing coordinates of bad pixels and column "type" defining bad pixel type ("hot", "bright", "dim", "dead").}
}
\value{
Data frame containing bad pixel coordinates and type, along with ID and size of clusters and superclusters to which each is assigned, and a shape classification based on these.
}
\description{
Identify clusters of adjacent pixels of each type in a bad pixel map, and also superclusters of adjacent hot and bright pixels.
}
\examples{
zz <- cluster.px(bp)
}

