% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_All-the-functions.R
\name{combine.maps}
\alias{combine.maps}
\title{Combine two bad pixel maps}
\usage{
combine.maps(new.map, old.map)
}
\arguments{
\item{new.map}{Data frame (as created using \link{\code{reset.bp}}) containing coordinates and types of pixels.}

\item{old.map}{Data frame (eg. official bad pixel map) containing coordinates and types of pixels.}
}
\value{
Data frame containing row and column coordinates of bad pixels, factor giving type of bad pixel, and factor noting which map the pixel was found in (old, new, or both)
}
\description{
Merge two data frames of bad pixels into one single list of coordinates, with source map and (where possible) error type marked.
}
\details{
NB. Currently only set up to compare 'official' (old) pixel map with new one. Should update this.
}
\examples{
matched <- combine.maps(bp, bpm)
}

