% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Line-detection.R
\name{convolve.lines}
\alias{convolve.lines}
\title{Identify bad rows or columns}
\usage{
convolve.lines(im, k.size = 5, horizontal = F, dim = F)
}
\arguments{
\item{im}{2d image array to be convolved.}

\item{k.size}{Size of (square) kernel to use. Default is 5.}

\item{horizontal}{Boolean: look for horizontal lines (T) instead of vertical (F)? Default is F, look for vertical lines.}

\item{dim}{Boolean: look for lines of dim pixels (T) instead of bright (F)? Default is F, enhance bright lines.}
}
\value{
Convolved image.
}
\description{
Perform a 2d convolution over an image to enhance lines of bright or dim pixels.
}
\examples{
zz <- convolve.lines(pw.m[,,"black", "160430"], k.size = 5)

}

