% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feature-identification.R
\name{count.clusters}
\alias{count.clusters}
\title{Count clusters by size}
\usage{
count.clusters(px, check.dir = F)
}
\arguments{
\item{px}{Data frame containing pixel coordinates to be examined}

\item{check.dir}{Boolean: instead of producing summary table, should paired t-test of mean dimensions of clusters be performed? Default is F.}
}
\value{
Data.frame summarising sizes of clusters found by dominant direction of spread
}
\description{
Count all clusters of given size in pixel map. Also checks main direction of spread (horizontal or vertical)
}

