% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feature-identification.R
\name{count.columns}
\alias{count.columns}
\title{Summarise columns by width}
\usage{
count.columns(px, midline = 1024.5)
}
\arguments{
\item{px}{Data frame containing pixel coordinates to be grouped}

\item{midline}{Location of panel midline, which electrically separates the upper and lower panels, and so cannot be crossed by column defects. Default is 1024.5; if the detector has no midline separation, enter NA.}
}
\value{
data.frame summarising multi-column features identified by panel, location and width
}
\description{
Group all adjacent column defects in pixel map and summarise each column defect by width
}

