% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pre-processing.R
\name{count.images}
\alias{count.images}
\title{Count available images}
\usage{
count.images(fpath = "./Image-data/")
}
\arguments{
\item{fpath}{Path to top level of stored images. Default is "./Image-data/"}
}
\value{
Data frame containing counts of .tif images in channel subfolders for each date, with image acquisition dates as row names.
}
\description{
Searches through folder structure to count number of available .tif images. Assumes files are organised as /Image-data/date/channel/image.tif
}
\examples{
image.count <- count.images()


}

