% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_All-the-functions.R
\name{dark.lines}
\alias{dark.lines}
\title{Find dark lines in an image}
\usage{
dark.lines(bright.image, dark.limit = 25000, min.length = 5,
  midline = 1024.5)
}
\arguments{
\item{bright.image}{Bright image array in which dark pixels are to be identified}

\item{dark.limit}{Integer grey-value below which a pixel will be classified as dark. Default is 25000}

\item{min.length}{Integer minimum length of a line segment to be considered a line (rather than a cluster). Default is 5}

\item{midline}{Y-coordinate of horizontal midline across panel, where upper and lower subpanels are electrically separated. Default is 1024.5; if no midline exists, use midline = NA}
}
\value{
Data frame of x and y coordinates of identified dark lines, with line type and feature type
}
\description{
Identify pixels below a certain threshold and clump with adjacent dark pixels to find long columns of dark pixels. (To find rows, the bright image should be transposed)
}
\examples{
dc <- dark.lines(pw.m[,,"white", "130613"])         # look for dark columns
dr <- dark.lines(t(pw.m[,,"white", "130613"]))      # look for dark rows

}

