% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classification.R
\name{dead.px}
\alias{dead.px}
\title{Identify dead pixels}
\usage{
dead.px(dt)
}
\arguments{
\item{dt}{Integer or string (format yymmdd) giving date to identify hot pxels in.}
}
\value{
Data frame containing coordinates of dead pixels, with a factor indicating the classification
}
\description{
Identify dead pixels in an image (any pixel with 0 value in white image)
}
\details{
Requires existence of image array \code{pw.m}
}

