% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_All-the-functions.R
\name{draw.outlines}
\alias{draw.outlines}
\title{Draw outlines of convex hulls}
\usage{
draw.outlines(px, im.dim = c(2048, 2048), ...)
}
\arguments{
\item{px}{Matrix or data frame of x and y coordinates to be clumped and outlined}
}
\description{
Given a set of coordinates, identify each cluster of adjacent pixels and draw the convex hull.
}

