% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_All-the-functions.R
\name{draw.panels.2048}
\alias{draw.panels.2048}
\title{Draw subpanels on 2048 x 2048 array}
\usage{
draw.panels.2048(...)
}
\arguments{
\item{...}{Optional graphical arguments}
}
\description{
Marks locations of 32 subpanels of 128 x 1024 pixels on plot
}

