% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_All-the-functions.R
\name{draw.panels}
\alias{draw.panels}
\title{Add panel edges to plot}
\usage{
draw.panels(p = panel.edges(), ...)
}
\arguments{
\item{p}{List containing y, a named vector of y-coordinates of panel corners, and x, a named vector of x-coordinates of panel corners.}
}
\description{
Mark edges of individual panels in array
}
\details{
Each detector panel outputs 128 x 1024 pixels, with some cropping at each edge of the combined image.
}
\examples{
pixel.image(b.150828)
draw.panels()            # marks specified panel boundaries


}

