% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classification.R
\name{edge.px}
\alias{edge.px}
\title{Return coordinates of edge pixels}
\usage{
edge.px(im, edge.width = 10)
}
\arguments{
\item{im}{Single-layer image array (needed to obtain dimensions)}

\item{edge.width}{Integer: number of pixels to treat as panel edges. Default is 10.}
}
\value{
Data frame containing coordinates of edge pixels
}
\description{
For a given image array, return the coordinates of pixels lying within a given distance of the edge
}
\details{
Edge pixels are applied to the outer edge of the image array, regardless of padding.
}

