% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{enhance.spots}
\alias{enhance.spots}
\title{Enlarge screen spots to include any adjacent dim pixels}
\usage{
enhance.spots(bpm)
}
\arguments{
\item{bpm}{Bad pixel map with categories assigned to each pair of coordinates, as generated by \link{\code{bad.px}}}
}
\value{
Updated ad pxiel map
}
\description{
Extend any screen spots in an image to include any adjacent dim pixels detected.
}

