% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feature-identification.R
\name{find.clusters}
\alias{find.clusters}
\title{Find and label clusters of pixels}
\usage{
find.clusters(px, midline = 1024.5)
}
\arguments{
\item{px}{Data frame containing pixel coordinates to be examined}

\item{midline}{Location of panel midline, which electrically separates the upper and lower panels, and so cannot be crossed by column defects. Default is 1024.5; if the detector has no midline separation, enter NA.}
}
\value{
Original pixel list, with column 'f.type' populated with factor 'cl.body' or 'cl.root' where appropriate.
}
\description{
Cluster pixels and identify them, labelling outer pixels as cluster bodies and the brightest central pixel as the root.
}

