% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feature-identification.R
\name{find.columns}
\alias{find.columns}
\title{Find column features in pixel map}
\usage{
find.columns(px, kernel.size = 5, midline = 1024.5, min.length = 3 *
  kernel.size, max.sep = 100)
}
\arguments{
\item{px}{Data frame containing pixel coordinates to be grouped}

\item{kernel.size}{Edge length of square kernel to be used in convolution. Default is 5 (aims to pick up line segments of length 5 or greater)}

\item{midline}{Location of panel midline, which electrically separates the upper and lower panels, and so cannot be crossed by column defects. Default is 1024.5; if the detector has no midline separation, enter NA.}

\item{min.length}{Minimum length of segment to retain after convolution. Default is 3 * kernel size.}
}
\value{
Data frame px, with column 'f.type' labelling columns found as feature type 'line.c'
}
\description{
Applies a linear convolution filter to identify vertical linear features in the pixel map.
}

