% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Line-detection.R
\name{find.lines}
\alias{find.lines}
\title{Find lines in image}
\usage{
find.lines(im, k.size = 5, threshold.at = 5500, sm.size = 11,
  min.length = 6, midline = 992.5, filter.at = list(cover = 0.5, filled =
  20), horizontal = F, dim.lines = F)
}
\arguments{
\item{im}{2d image array to be convolved.}

\item{k.size}{Size of (square) kernel to use. Default is 5.}

\item{threshold.at}{Numeric value at which to threshold convolved data. Default is 5500, which will highlight rows that are approximately 300 grey values higher than both of their neighbours in the direction of interest.}

\item{sm.size}{Length of (linear) kernel to use. Default is 11.}

\item{min.length}{Integer: minimum number of pixels to accept as a line after smoothing.}

\item{horizontal}{Boolean: look for horizontal lines (T) instead of vertical (F)? Default is F, look for vertical lines.}

\item{dim.lines}{Boolean: look for lines of dim pixels (T) instead of bright (F)? Default is F, enhance bright lines.}
}
\value{
Image array with lines marked
}
\description{
Use a combination of convolution & thresholding to identify consistently bright or dim columns or rows in an image
}

