% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feature-identification.R
\name{find.rows}
\alias{find.rows}
\title{Find row features in pixel map}
\usage{
find.rows(px, kernel.size = 5, min.length = 3 * kernel.size)
}
\arguments{
\item{px}{Data frame containing pixel coordinates to be grouped}

\item{kernel.size}{Edge length of square kernel to be used in convolution. Default is 5 (aims to pick up line segments of length 5 or greater)}

\item{min.length}{Minimum length of segment to retain after convolution. Default is 3 * kernel size.}
}
\value{
Data frame px, with column 'f.type' labelling columns found as feature type 'line.c'
}
\description{
Applies a linear convolution filter to identify horizontal linear features in the pixel map.
}

