% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-fitting.R
\name{fit.circular.lm.poly}
\alias{fit.circular.lm.poly}
\title{Linear regression over distance from centre, including polynomial terms of z}
\usage{
fit.circular.lm.poly(image, o = 2, centre = c(1023.5, 992.5))
}
\arguments{
\item{image}{Single-layer (1996x1996) array image to be fitted.}

\item{o}{Order of polynomial to include. Default is 2}

\item{centre}{Vector of coordinates of centre of spot. Default is centre of uncropped detector, c(1023.5, 992.5)}
}
\value{
Linear model fitted to given image
}
\description{
Fit a circular spot model with linear gradients to the image, including polynomial terms of z as explanatory variables
}
\examples{
circ.poly.lm <- fit.circular.lm.poly(pw.m)
circ.poly.res <- matrix(circ.poly.lm$residuals, ncol = 1996)
pixel.image(circ.poly.res)

}

