% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-fitting.R
\name{fit.circular.lm}
\alias{fit.circular.lm}
\title{Linear regression over distance from centre}
\usage{
fit.circular.lm(image, centre = c(1023.5, 992.5))
}
\arguments{
\item{image}{Single-layer (1996x1996) array image to be fitted.}

\item{centre}{Vector of coordinates of centre of spot. Default is centre of uncropped detector, c(1023.5, 992.5)}
}
\value{
Linear model fitted to given image
}
\description{
Fit a circular spot model with linear gradients to the image.
}
\examples{
circ.lm <- fit.circular.lm(pw.m)
circ.res <- matrix(circ.lm$residuals, ncol = 1996)
pixel.image(circ.res)
}

