% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-fitting.R
\name{fit.panel.lm.poly}
\alias{fit.panel.lm.poly}
\title{Per-panel linear regression, with polynomial terms}
\usage{
fit.panel.lm.poly(image, o = 2)
}
\arguments{
\item{image}{Single-layer array image (1996x1996) to be fitted.}

\item{o}{Order of polynomial to include. Default is 2}
}
\value{
List containing a matrix of fitted values, and a matrix of model coefficients for each panel.
}
\description{
Fit a linear gradient over each of the 32 subpanels within the image, including polynomial terms of x and y.
}
\examples{
panel.lm.2 <- fit.panel.lm.poly(circ.res.2, o = 2)
panel.res.2 <- circ.res.2 - panel.lm.2$fitted.values
pixel.image(panel.res.2)
}

