% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-fitting.R
\name{fit.panel.lm}
\alias{fit.panel.lm}
\title{Per-panel linear regression}
\usage{
fit.panel.lm(image)
}
\arguments{
\item{image}{Single-layer array image (1996x1996) to be fitted.}
}
\value{
List containing a matrix of fitted values, and a matrix of model coefficients for each panel.
}
\description{
Fit a linear gradient over each of the 32 subpanels within the image
}
\examples{
panel.lm <- fit.panel.lm(circ.res)
panel.res <- circ.res - panel.lm$fitted.values
pixel.image(panel.res)
}

