% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-fitting.R
\name{fit.panels}
\alias{fit.panels}
\title{Return fitted values from panelwise coefficients}
\usage{
fit.panels(coeffs)
}
\arguments{
\item{coeffs}{32 x 3 matrix of coefficients (offset, x, y) for each panel}
}
\value{
List containing 1996x1996 matrix of per-panel offset values and 1996x1996 matrix of per-panel gradient values.
}
\description{
Given a matrix of coefficients such as from \link{\code{fit.panel.lm}}, return a list containing values fitted to the panel coordinates.
}
\examples{
panel.lm <- fit.panel.lm(circ.res)
zz <- fit.panels(panel.lm$models)
pixel.image(zz$grad)
}

