% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary-operations.R
\name{flat.field.corrected}
\alias{flat.field.corrected}
\title{Flat-field/offset correction of image}
\usage{
flat.field.corrected(dt)
}
\arguments{
\item{dt}{Integer/string: date to create flat-field image}
}
\value{
1996x1996 array containing corrected image
}
\description{
Use full set of black, white and grey images to obtain offset-corrected image. Requires array \code{pw.m}, as generated by \link{\code{load.pixel.means()}}
}
\examples{
corr <- flat.field.corrected(160314)
}

