% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{focal.plot}
\alias{focal.plot}
\title{Plot focal area}
\usage{
focal.plot(im, centre, surround = 5, dp = 1, scale.by = 1000,
  lbl.cex = 0.7, bad.px, bpx.cex = 2.5, ...)
}
\arguments{
\item{im}{2d image array to plot}

\item{centre}{Vector of x and y coordinates of pixel of particular interest}

\item{surround}{Integer: set size of area surrounding pixel of interest to be displayed. Default is 5, which displays an 11x11 square centred on the pixel of interest.}

\item{dp}{Integer: display pixel values to how many decimal places? Default is 1.}

\item{scale.by}{Integer: divide pixel values by this number for easier display and comparison. Default is 1000, so a value of 65535 will be displayed as 65.5}

\item{lbl.cex}{Magnification factor to apply to labels in each cell. Default is 0.7}

\item{bad.px}{Optional bad pixel map to be used to highlight any pixels in the display area that have been identified as defective.}

\item{bpx.cex}{Magnification factor to apply to symbol used to highlight bad pixels. Default is 2.5}

\item{...}{Additional optional graphical arguments}
}
\description{
Create pixel image of area surrounding a point of interest. Pixel values are labelled and bad pixels identified highlighted
}

