% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parametric-modelling.R
\name{gaussian.spot.ls}
\alias{gaussian.spot.ls}
\title{Least Squares fit of bivariate Gaussian surface to image}
\usage{
gaussian.spot.ls(im, par, x0.target = c(0, ncol(im)), y0.target = c(0,
  nrow(im)))
}
\arguments{
\item{im}{2d image array to which surface is to be fitted.}

\item{par}{Starting values for the parameters to be fitted.}

\item{x0.target}{Range of values within which the x-origin will be constrained to lie. Default is image area.}

\item{y0.target}{Range of values within which the y-origin will be constrained to lie. Default is image area.}
}
\value{
List of fitted parameters and convergence statistics produced by \link{\code{optim}}.
}
\description{
Use \link{\code{optim}} to fit an elliptical bivariate Gaussian surface to a two-dimensional image by minimizing sum of squared errors.
}
\examples{
spot <- gaussian.spot.ls(pw.m[,,"grey", "160430"], c(A = 15000, x0 = 1024.5, y0 = 1024.5, sig.x = 500, sig.y = 500))

}

