% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_All-the-functions.R
\name{gaussian.spot.mat}
\alias{gaussian.spot.mat}
\title{Produce bivariate density surface}
\usage{
gaussian.spot.mat(par, arr.dim = c(2048, 2048))
}
\arguments{
\item{par}{Vector of named parameters}

\item{arr.dim}{Dimensions of array to be produced. Default is c(2048, 2048).}
}
\value{
Array of given dimensions containing evaluated surface at each point.
}
\description{
Produce bivariate density surface from given parameters and return as array
}
\examples{
fitted.spot <- gaussian.spot.ls(pw.m[,,"grey", "160430"], c(A = 15000, x0 = 1024.5, y0 = 1024.5, sig.x = 500, sig.y = 500))
pixel.image(gaussian.spot.mat(fitted.spot$par, arr.dim = c(2048, 2048)))

}

