% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_All-the-functions.R
\name{get.bad.pixels}
\alias{get.bad.pixels}
\title{Identify extreme values in image}
\usage{
get.bad.pixels(image, method = "sd", n.sigma = 1.96, span = 1/15,
  display = T, panels = T)
}
\arguments{
\item{image}{Single-layer 1996x1996 image array}

\item{method}{String: either "SD" or "NMAD". Threshold will be set according to multiples of SD or NMAD of columnwise residuals.}

\item{n.sigma}{Numeric scalar: how many multiples of the SD/NMAD will be used as threshold for 'extreme' values.}

\item{span}{Numeric value: smoothing span to be used in Loess spline}

\item{display}{Boolean: plot the results?}

\item{panels}{Boolean: if plotting the results, show subpanel edges?}
}
\description{
Find columnwise extreme values in image array
}
\details{
For each column of each panel, use Loess smoothing to find central spline and calculate residuals. Any points with a residual above a certain threshold are flagged as bad pixels.
}

