% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classification.R
\name{get.dim.bright.px}
\alias{get.dim.bright.px}
\title{Identify bright and dim pixels}
\usage{
get.dim.bright.px(im)
}
\arguments{
\item{im}{Two-dimensional image array containing pixel values to be classified}

\item{cropped}{Boolean: crop edge pixels before thresholding? Default is T.}
}
\value{
Data frame containing coordinates of bright and dim pixels, with a factor indicating the classification
}
\description{
Use the thresholds returned by \link{\code{bad.px.limits}} to classify bright and dim pixels in an image.
}
\examples{
bp <- get.dim.bright.px(res[,,"white", "160314"])
table(bp$type)

}

