% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{get.focus}
\alias{get.focus}
\title{Get coordinates of points to plot zoom}
\usage{
get.focus(points, surround = 3)
}
\arguments{
\item{points}{Matrix or data frame, the first two columns of which are the x and y coordinates of the points to be plotted.}

\item{surround}{Number of additional pixels to plot in every direction around the selected points}
}
\value{
matrix of coordinates to plot
}
\description{
Given a set of points and the surrounding area to plot, obtain coordinates to create a plot of a small subset of an image
}

