% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_All-the-functions.R
\name{get.transitions}
\alias{get.transitions}
\title{Get transition matrices for bad pixel maps}
\usage{
get.transitions(bp)
}
\arguments{
\item{bpx}{List of dataframes containing bad pixel coordinates and types}
}
\value{
List of tables showing movement between states from one acquisition to the next
}
\description{
Given a list of bad pixel maps in coordinate form, return a list of transition matrices showing movements between states observed between one acquisition and the next.
}

