% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_All-the-functions.R
\name{he.spot.lm}
\alias{he.spot.lm}
\title{Fit hyper- or hypo-elliptical spot model}
\usage{
he.spot.lm(image, n = 2, order = 2, centre = c(1023.5, 992.5),
  robust = F)
}
\arguments{
\item{image}{Single-layer (1996x1996) array image to be fitted.}

\item{n}{Power to which x- and y- distances are to be raised to fit model.}

\item{order}{Order of polynomial to include. Default is 2}

\item{centre}{Vector of coordinates of centre of spot. Default is centre of uncropped detector, c(1023.5, 992.5)}

\item{robust}{Boolean: use robust model fitting (\link{\code{rlm}})? Default is F (use \link{\code{lm}} for regression).}
}
\value{
Linear model fitted to given image
}
\description{
Uses linear regression over x- and y-distance from centre to the power n to fit a hyper- or hypo-elliptical spot
}
\details{
1 < n < 2: rhombus with convex edges; n = 2: an ellipse; n > 2: rectangle with rounded corners.
}
\examples{
spot <- spot.lm(pw.m[,,"white", "150828"], n = 2, robust = T)
spot.res <- matrix(spot$residuals, ncol = 1996)
pixel.image(spot.res)
}

