% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classification.R
\name{hot.px}
\alias{hot.px}
\title{Identify hot pixels}
\usage{
hot.px(dt)
}
\arguments{
\item{dt}{Integer or string (format yymmdd) giving date to identify hot pxels in.}
}
\value{
Data frame containing coordinates of hot pixels, with a factor indicating the classification
}
\description{
Identify hot pixels in an image (any pixel with maximum possible value in black image)
}
\details{
Requires existence of image array \code{pw.m}
}

