% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import.R
\name{import.acq}
\alias{import.acq}
\title{Import a single day's acquisition into a 2048 x 2048 array}
\usage{
import.acq(acq.folder, subfolders = c("black", "grey", "white"),
  panel.dim = c(2048, 2048))
}
\description{
Import all images for one acquisition date into a 2048 x 2048 array, with layers corresponding to different power settings. Cropped pixels are padded with NA.
}
\examples{
im.160430 <- import.acq("/home/clair/Documents/Pixels/Image-data/160430")

}

