% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_All-the-functions.R
\name{iqr.levels}
\alias{iqr.levels}
\title{Set contour plot levels}
\usage{
iqr.levels(data)
}
\arguments{
\item{data}{Matrix of data to plot}
}
\value{
Vector of calculated levels
}
\description{
Support function: returns a vector of levels to be used in contour plotting, cutting data at outlier (lq - 1.5*IQR), lq, 3 even bins between lq and median, median, 3 even bins between median and uq, uq, and outlier (1.5*iqr + uq).
}
\examples{
    filled.contour(x = c(1:dim(data)[1]), y = c(1:dim(data)[2]),
                   data,
                   levels = iqr.levels(data))


}

