% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{Johnson.QQ}
\alias{Johnson.QQ}
\title{Johnson Q-Q plot}
\usage{
Johnson.QQ(data, quantiles = c(1:999)/1000, grid.quantiles = c(0.01, 0.99),
  title = "Johnson Q-Q plot", ...)
}
\arguments{
\item{data}{Image array of 1996 x 1996 pixels}

\item{quantiles}{Vector of quantiles to plot; default is c(1:999)/1000}

\item{grid.quantiles}{Vector of quantiles of particular interest, to mark with gridlines}

\item{title}{String containing plot title. Default is "Johnson Q-Q plot".}

\item{...}{Additional plotting arguments}
}
\description{
Produce a Q-Q plot of data quantiles vs Johnson distribution quantiles, with quantiles of interest marked.
}
\examples{
Johnson.QQ(pw.sd[,,"black", "150828"])
}

