% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_All-the-functions.R
\name{join.panels}
\alias{join.panels}
\title{Merge subpanels into image array}
\usage{
join.panels(data, left.pad = 2, upper.pad = 20)
}
\arguments{
\item{data}{128x1024x32 array of subpanel images}

\item{left.pad}{Integer numer of pixels cropped from left-hand edge of image. Default is 2}

\item{upper.pad}{Integer numer of pixels cropped from upper edge of image. Default is 20}
}
\description{
Take a subpanel image array and recombine into single panel image
}
\examples{

zz <- split.panels(pw.m)
qq <- join.panels(zz)
all(pw.m == qq)               # TRUE
}

