% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feature-identification.R
\name{label.screen.spots}
\alias{label.screen.spots}
\title{Label identified screen spots in pixel list}
\usage{
label.screen.spots(px, spot.list)
}
\arguments{
\item{px}{Data frame containing pixel coordinates to be grouped}

\item{spot.list}{List of matrices of coordinates of identified screen spots, as returned by \link{\code{screen.spots}}}
}
\value{
Original pixel list, with column 'f.type' now showing identified screen spots
}
\description{
Given a pixel list and list of identified screen spots, return a pixel list with defects belonging to screen spots identified.
}

