% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pre-processing.R
\name{load.daily}
\alias{load.daily}
\title{Load Daily Images}
\usage{
load.daily(img.date, channel, x = 1996, y = 1996, z = 20, progress = F,
  fpath = "./Image-data/")
}
\arguments{
\item{img.date}{Date of images to import, in format yymmdd.}

\item{channel}{Specify channel to import: black, grey or white.}

\item{x}{Width of image. Default is 1996.}

\item{y}{Height of image. Default is 1996.}

\item{z}{Number of images to import. Default is 20.}

\item{progress}{Display progress bar? T or F (default is F)}

\item{fpath}{Path to top level of stored images. Default is "./Image-data/"}
}
\description{
Import one day's images into a single array
}
\examples{
daily <- load.daily(150828, "black")


}

