% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import.R
\name{load.images}
\alias{load.images}
\title{Load one day's acquisitions from one batch}
\usage{
load.images(img.date, batch,
  fpath = "/home/clair/Documents/Pixels/Image-data/", x = 1996, y = 1996,
  z = 20)
}
\arguments{
\item{img.date}{Date of images to import, in format yymmdd.}

\item{batch}{Specify batch to import: black, grey or white.}

\item{fpath}{Path to top level of stored images. Default is "/home/clair/Documents/Pixels/Image-data/"}

\item{x}{Width of image. Default is 1996.}

\item{y}{Height of image. Default is 1996.}

\item{z}{Number of images to import. Default is 20.}
}
\description{
Import one day's images from a single batch into an array. If the function is called by the user, the data is loaded into an object named automatically by the function. If called by another function, will return an array.
}
\examples{
load.images(150828, "black")   # will create array b.150828


}

