% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import.R
\name{load.objects}
\alias{load.objects}
\title{Create array from stored .rds objects}
\usage{
load.objects(fpath = "./02_Objects/images", otype = "pwm", acq.list)
}
\arguments{
\item{fpath}{Path to top level of stored .rds objects. Default is "./02_Objects/images"}

\item{otype}{String specifying which type of image should be imported. Default is "pwm" (pixelwise mean images).}

\item{acq.list}{Vector containing acquisitions to import. If left blank, will import every file in the target folder named "<otype>-xxxxxx.rds".}
}
\description{
Load all .rds ojects of a specified type into a single labelled array.
}
\details{
RDS objects for import can be downloaded from https://github.com/ClairBee/Pixels/tree/master/02_Objects
}

