% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_All-the-functions.R
\name{load.pixel.maps}
\alias{load.pixel.maps}
\title{Quickly load files}
\usage{
load.pixel.maps(fpath = "./Other-data/")
}
\arguments{
\item{fpath}{Path to top level of stored images. Default is "./Other-data/"}
}
\description{
Load all pixelwise summaries and bad pixel map in one step.
}
\details{
Imports pixelwise mean and SD for all files, and bad pixel map
}
\examples{
load.pixel.maps()

}

