% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import.R
\name{load.pixel.means.2048}
\alias{load.pixel.means.2048}
\title{Load pixel means into 2048 x 2048 array}
\usage{
load.pixel.means.2048(fpath = "./02_Objects/images")
}
\description{
Load all pixelwise mean summaries into a single labelled array of size 2048 x 2048 (detector size). Cropped pixels are labelled as NA.
}

