% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import.R
\name{load.pixel.means}
\alias{load.pixel.means}
\title{Create array of all pixelwise means}
\usage{
load.pixel.means(acq.list, fpath = "./02_Objects/images")
}
\arguments{
\item{acq.list}{Vector containing acquisitions to import. If left blank, will import every file in the target folder named "pwm-xxxxxx.rds".}

\item{fpath}{Path to top level of stored .rds objects. Default is "./02_Objects/images"}
}
\description{
Load all pixelwise mean summaries (already created and stored as .rds objects) into a single labelled array.
}

