% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import.R
\name{load.pixel.sds}
\alias{load.pixel.sds}
\title{Create array of all pixelwise SDs}
\usage{
load.pixel.sds(acq.list, fpath = "./02_Objects/stdevs")
}
\arguments{
\item{acq.list}{Vector containing acquisitions to import. If left blank, will import every file in the target folder named "pwm-xxxxxx.rds".}

\item{fpath}{Path to top level of stored .rds objects. Default is "./02_Objects/images"}
}
\description{
Load all pixelwise standard deviation matrices into a single labelled array.
}

