% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import.R
\name{load.profiles}
\alias{load.profiles}
\title{Load xml profiles of one day's acquisitions from one channel}
\usage{
load.profiles(img.date, channel, fpath = "./Image-data/")
}
\arguments{
\item{img.date}{Date of images to import, in format yymmdd.}

\item{channel}{Specify channel to import: black, grey or white.}

\item{fpath}{Path to top level of stored images. Default is "./Image-data/"}
}
\description{
Support function: import xml profiles of one day's images into a single data frame
}
\examples{
b.150828.xml <- load.profiles(150828, "black")


}

