% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classification.R
\name{locally.bright.px}
\alias{locally.bright.px}
\title{Identify locally bright pixels}
\usage{
locally.bright.px(dt, setting)
}
\arguments{
\item{dt}{Integer or string (format yymmdd) giving date to identify bright pixels in.}
}
\value{
Data frame containing coordinates of dead pixels, with a factor indicating the classification
}
\description{
Identify locally bright pixels in an image
}
\details{
Requires existence of image array \code{pw.m} and median-differenced array \code{md}
}

