% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_All-the-functions.R
\name{locally.dim.px}
\alias{locally.dim.px}
\title{Identify locally dim pixels}
\usage{
locally.dim.px(dt, setting)
}
\arguments{
\item{dt}{Integer or string (format yymmdd) giving date to identify dim pixels in.}
}
\value{
Data frame containing coordinates of dead pixels, with a factor indicating the classification
}
\description{
Identify locally dim pixels in an image
}
\details{
Requires existence of image array \code{pw.m} and median-differenced array \code{md}
}

