% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-fitting.R
\name{lowess.per.column}
\alias{lowess.per.column}
\title{Per-column loess smoothing}
\usage{
lowess.per.column(image, span = 1/15, midline = 992.5)
}
\arguments{
\item{image}{Single-layer array image (1996x1996) to be fitted.}

\item{span}{Span to be used for loess smoothing. Default is 1/15.}
}
\value{
1996x1996 matrix of smoothed values.
}
\description{
Fit a loess spline to each column of the image using \link{\code{lowess}}. The upper and lower detector panels are treated separately.
}
\examples{
smoothed <- lowess.per.column(panel.res)
smoothed.res <- panel.res - smoothed
pixel.image(smoothed.res)
s.hist(smoothed.res)
}

