% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc-functions.R, R/zz_All-the-functions.R
\name{m2r}
\alias{m2r}
\title{Convert matrix to raster}
\usage{
m2r(im)

m2r(im)
}
\arguments{
\item{im}{Single-layer image array to be converted to a raster}

\item{im}{Single-layer image array to be converted to a raster}
}
\value{
RasterLayer object containing the image values

RasterLayer object containing the image values
}
\description{
Convert an image array to a raster file, maintaining the correct orientation for plotting

Convert an image array to a raster file, maintaining the correct orientation for plotting
}

