% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{med.diffs}
\alias{med.diffs}
\title{Get median differences over image}
\usage{
med.diffs(im)
}
\arguments{
\item{im}{Single-layer image array}
}
\value{
Image array containing pixelwise differences
}
\description{
Run a median filter over an image to obtain a smoothed version, then subtract this from the original image to highlight any points that are unusually high in relation to their neighbours
}

