% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-fitting.R
\name{minipanel.lm}
\alias{minipanel.lm}
\title{Mini-panel linear regression}
\usage{
minipanel.lm(image, terms = "x + y", robust = F, left.pad = 2,
  upper.pad = 20)
}
\arguments{
\item{image}{Single-layer array image (1996x1996) to be fitted.}

\item{terms}{String specifying terms to be fitted to each subpanel. Default is "x + y", fitting a linear model to the x and y coordinates of each subpanel without considering interactions.}

\item{robust}{Boolean: use robust regression (\link{\code{rlm}})? Default is F (use \link{\code{lm}} for regression).}

\item{left.pad}{Integer value: how many pixels are cropped from left image edge? Default is 2.}

\item{upper.pad}{Integer value: how many pixels are cropped from upper image edge? Default is 20.}
}
\value{
List containing the terms applied, a matrix of fitted values, and a matrix of model coefficients for each panel.
}
\description{
Fit a linear gradient over each of the 4 minipanels within each of the 32 subpanels within the image
}
\examples{
panel.lm <- panel.lm(circ.res, "y")                 # finds coefficients of intercept and y.
panel.res <- circ.res - panel.lm$fitted.values
minipanel <- minipanel.lm(panel.res, terms = "x + y", robust = T)
res <- panel.res - minipanel$fitted.values
}

