% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{modal.density}
\alias{modal.density}
\title{Modal density}
\usage{
modal.density(dat)
}
\arguments{
\item{dat}{Data for which bounds are to be calculated}
}
\value{
Value at which modal density occurs
}
\description{
Calculate location of modal density of a distribution
}

