% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{o.plot}
\alias{o.plot}
\title{Vector overlay plot}
\usage{
o.plot(data, add = F, ...)
}
\arguments{
\item{data}{Vector of values to plot (eg. column of pixel values)}

\item{add}{Boolean: add values to current plot, or start a new plot? Default is F.}
}
\description{
Base plotting function, specifying type "o" (points plotted over lines), pch = 20, cex = 0.7. Other plotting parameters can be passed as arguments as usual.
}
\examples{
o.plot(b.150828[4,,1])
o.plot(b.150828[5,,1], col = adjustcolor("red", alpha = 0.2), add = T)


}

