% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Line-detection.R
\name{offset.lines}
\alias{offset.lines}
\title{Find dim or bright lines in an image}
\usage{
offset.lines(md, k.size = 5, threshold.at = 5500, min.length = 10,
  midline = 1024.5)
}
\arguments{
\item{k.size}{Size of (square) kernel to use. Default is 5.}

\item{threshold.at}{Numeric value at which to threshold convolved data. Default is 5500, which will highlight rows that are approximately 300 grey values higher than both of their neighbours in the direction of interest.}

\item{min.length}{Integer: minimum number of pixels to accept as a line segment after smoothing. Default is 10.}

\item{midline}{Y-coordinate of horizontal midline across panel, where upper and lower subpanels are electrically separated. Default is 1024.5; if no midline exists, use midline = NA}
}
\value{
Data frame containing x and y coordinates of identified lines, and classification of each segment as either a line root, line body, or channel offset.
}
\description{
Use convolution and filtering to identify vertical segments of bright or dim pixels.
}
\examples{

}

