% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_All-the-functions.R
\name{panel.edges}
\alias{panel.edges}
\title{Get coordinates of panel edges}
\usage{
panel.edges(left.crop = 0, upper.crop = 0, width = 128, height = 1024,
  x.dim = 2048, y.dim = 2048)
}
\arguments{
\item{left.crop}{Number of pixels cropped from left-hand edge of image. Default is 2.}

\item{upper.crop}{Number of pixels cropped from lower edge of image. Default is 20.}

\item{width}{Width of each panel, in pixels. Default is 128.}

\item{height}{Height of each panel, in pixels. Default is 1024.}

\item{x.dim}{Width of image returned, in pixels. Default is 1996.}

\item{y.dim}{Height of image returned, in pixels. Default is 1996.}
}
\description{
Identify coordinates of individual detector panels, given their dimensions and those of the whole sensor.
}
\details{
Each detector panel outputs 128 x 1024 pixels, with some cropping at each edge of the combined image.
}
\examples{
panels <- panel.edges()


}

