% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-fitting.R
\name{panel.lm}
\alias{panel.lm}
\title{Per-panel linear regression}
\usage{
panel.lm(image, terms = "x + y", robust = F)
}
\arguments{
\item{image}{Single-layer array image (1996x1996) to be fitted.}

\item{terms}{String specifying terms to be fitted to each subpanel. Default is "x + y", fitting a linear model to the x and y coordinates of each subpanel without considering interactions.}

\item{robust}{Boolean: use robust model fitting (\link{\code{rlm}})? Default is F (use \link{\code{lm}} for regression).}
}
\value{
List containing the terms applied, a matrix of fitted values, and a matrix of model coefficients for each panel.
}
\description{
Fit a linear gradient over each of the 32 subpanels within the image
}
\examples{
panel.lm <- panel.lm(circ.res, "poly(x, 2) + y")    # finds coefficients of intercept, x, x^2, y.
panel.lm <- panel.lm(circ.res, "y")                 # finds coefficients of intercept and y.
panel.res <- circ.res - panel.lm$fitted.values
pixel.image(panel.res)
}

