% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pre-processing.R
\name{pixel.contour}
\alias{pixel.contour}
\title{Contour plot of pixel values}
\usage{
pixel.contour(data, title = "", midpoint = "mean")
}
\arguments{
\item{data}{2-dimensional matrix containing values to be plotted}

\item{title}{String containing title to be printed with plot}

\item{midpoint}{Either 'mean' or 'median'. Default is 'mean'}
}
\description{
Create a filled contour plot with pixels shaded according to their distance from the mean value.
}
\details{
Pixels more than 2sd above the mean are red, pink or purple. Pixels within 2sd of the mean are orange, yellow or green. Pixels more than 2sd below the mean are blue.
}
\examples{
pixel.contour(pw.mean)


}

