% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pre-processing.R
\name{pixel.image}
\alias{pixel.image}
\title{Display pixel values as image}
\usage{
pixel.image(data, title = "", x.range = c(1:1996), y.range = c(1:1996),
  midpoint = "mean")
}
\arguments{
\item{data}{2-dimensional matrix containing values to be plotted}

\item{title}{String containing title to be printed with plot}

\item{x.range}{Vector range showing x-range of cells to be included in plot}

\item{y.range}{Vector range showing y-range of cells to be included in plot}

\item{midpoint}{Either 'mean' or 'median'. Default is 'mean'}
}
\description{
Create an image with pixels shaded according to their distance from the mean value. Seems to display at lower resolution than contour plot, but much quicker.
}
\details{
Pixels more than 2sd above the mean are red, pink or purple. Pixels within 2sd of the mean are orange, yellow or green. Pixels more than 2sd below the mean are blue.
}
\examples{
pixel.image(pw.mean)


}

