% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R, R/zz_All-the-functions.R
\name{pixel.plot}
\alias{pixel.plot}
\title{Plot given coordinates}
\usage{
pixel.plot(px, xlim = c(0, 2048), ylim = c(0, 2048), pch = 15,
  panels = F, cex = 0.4, main = "", xlab = "", ylab = "", ...)

pixel.plot(px, xlim = c(0, 2048), ylim = c(0, 2048), pch = 15,
  panels = F, cex = 0.4, main = "", xlab = "", ylab = "", ...)
}
\arguments{
\item{px}{Coordinates of pixels to be plotted}

\item{px}{Coordinates of pixels to be plotted}
}
\description{
Plot pixel map

Plot bad pixel map on 2048 x 2048 grid
}
\examples{
pixel.plot(which(pw.m[,,"white"] < 15000, arr.ind = T))

pixel.plot(which(pw.mean > 25000, arr.ind = T))


}

