% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_All-the-functions.R
\name{pixelwise.mean}
\alias{pixelwise.mean}
\title{Get pixelwise mean of daily acquisitions}
\usage{
pixelwise.mean(data)
}
\arguments{
\item{data}{Array with 3 or more dimensions}
}
\value{
Array containing pixelwise mean of all acquired values
}
\description{
Return pixelwise mean value across an array
}
\examples{
pw.mean <- pixelwise.mean(daily)


}

