% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary-operations.R
\name{pixelwise.sd}
\alias{pixelwise.sd}
\title{Get pixelwise standard deviation of daily acquisitions}
\usage{
pixelwise.sd(data)
}
\arguments{
\item{data}{Array with 3 or more dimensions}
}
\value{
Array containing pixelwise standard deviation of all acquired values
}
\description{
Return pixelwise SD across an array
}
\examples{
pw.sd <- pixelwise.sd(daily)


}

