% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{plot.map.comparison}
\alias{plot.map.comparison}
\title{Plot a combined pixel map}
\usage{
plot.map.comparison(map, img.date, colour = "all")
}
\arguments{
\item{map}{Data frame (as created using \link{\code{combine.maps}}) containing coordinates and sources of pixels.}

\item{img.date}{Integer or string giving acquisition date for which bad pixels are to be plotted.}

\item{colour}{Which image batch should be plotted? Default is "all", which will plot all three sequentially.}
}
\description{
Compare bad pixels identified by two different pixel maps by plotting their values.
}
\examples{
matched <- combine.maps(bp, bpm)
plot.map.comparison(matched)
}

