% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conversion.R
\name{px2ppp}
\alias{px2ppp}
\title{Convert pixel coordinates to PPP object}
\usage{
px2ppp(px, im.dim = c(2048, 2048), crop = list(c(1, im.dim[1]), c(1,
  im.dim[2])))
}
\arguments{
\item{px}{data.frame or matrix, the first two columns of which contain the X and Y coordinates of points to be marked on the raster}

\item{im.dim}{Vector of dimensions for the final raster. Default is c(2048, 2048).}

\item{crop}{List of two vectors of crop boundaries. Default is list(c(1, im.dim[1]), c(1, im.dim[2])), which crops to original image area.}
}
\value{
ppp object of the required dimensions
}
\description{
Convert coordinates from bad pixel map to a PPP object for spatial analysis. The resulting image can be further cropped if necessary.
}

