% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc-functions.R
\name{px2r}
\alias{px2r}
\title{Convert points to raster}
\usage{
px2r(px, im.dim = c(2048, 2048))
}
\arguments{
\item{px}{data.frame or matrix, the first two columns of which contain the X and Y coordinates of points to be marked on the raster}

\item{im.dim}{Vector of dimensions for the final raster. Default is c(2048, 2048).}
}
\value{
Raster object with the coordinates marked as 1, everything else as NA.
}
\description{
Support function: convert matrix of coordinates into a raster, so that convolution, clumping etc can be applied
}

