% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc-functions.R, R/zz_All-the-functions.R
\name{r2m}
\alias{r2m}
\title{Convert raster to matrix}
\usage{
r2m(rast)

r2m(rast)
}
\arguments{
\item{rast}{Single-layer RasterLayer object to be converted to a 2d array}

\item{rast}{Single-layer RasterLayer object to be converted to a 2d array}
}
\value{
Single-layer image array

Single-layer image array
}
\description{
Convert a raster to an image array, maintaining the correct orientation for plotting

Convert a raster to an image array, maintaining the correct orientation for plotting
}

