% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thresholding.R
\name{reset.bp}
\alias{reset.bp}
\title{Create map of 'absolute' bad pixels}
\usage{
reset.bp(img.dt)
}
\arguments{
\item{img.date}{Integer or string giving date to be}
}
\value{
Data frame containing coordinates and classifications of all absolutely defective pixels.
}
\description{
Create a data frame containing coordinates of 'hot' and 'dead' pixels in each image batch for a given date.
}
\details{
'Hot' pixels are always on, with a pixelwise mean value of 65535 across all 20 images. 
'Dead' pixels are always off, with a pixelwise mean value of 0 across all 20 images.

Requires objects \code{pw.b}, \code{pw.g} and \code{pw.g}, as created by \link{\code{load.pixel.maps}}.
}
\examples{
bp.160314 <- reset.bp(160314)
bp.160314 <- reset.bp("160314")          # gives same result

}

