% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{s.hist}
\alias{s.hist}
\title{Histogram plot with SD levels shown}
\usage{
s.hist(data, scale = sd.levels(data), scale.colours = sd.colours(), xlim,
  ...)
}
\arguments{
\item{data}{Array of numbers to be plotting in histogram}

\item{scale}{Vector of scale cutpoints. Default is \link{\code{sd.levels()}}.}

\item{scale.colours}{Vector of scale colours. Default it \link{\code{sd.lcolours()}}}

\item{xlim}{Vector of lower and upper x-limits. If not provided, will use central 95% of normal distribution with mean & sd of observed data.}
}
\description{
Producs a histogram with coloured bars at the bottom showing scale used. Defaults can be used to show legend for plots produced using \link{\code{pixel.image}}.
}
\examples{
pixel.image(b.150828)
scale.hist(b.150828)
}

