% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{sample.healthy}
\alias{sample.healthy}
\title{Create sample of 'healthy' pixels}
\usage{
sample.healthy(bad.pixels, n = nrow(bad.pixels))
}
\arguments{
\item{bad.pixels}{Data frame or matrix whose first two columns are X and Y coordinates of identified bad pixels.}

\item{n}{Size of sample to create. Default is to create a sample of same size as bad pixel list.}
}
\value{
Matrix containing X and Y coordinates of sampled pixels.
}
\description{
Given a matrix of coordinates of bad pixels, create a same-size random sample of healthy pixels that do not appear on the bad pixel list.
}
\examples{
bp.160314 <- reset.bp(160314)
samp.160314 <- sample.healthy(bp.160314)


}

