% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary-operations.R
\name{save.summary}
\alias{save.summary}
\title{Store summary statistics for an image batch in a data frame}
\usage{
save.summary(data)
}
\arguments{
\item{data}{Array with 3 or more dimensions}
}
\value{
Updates global data frame containing summary statistics for multiple images.
}
\description{
Populate a data frame with the summary statistics provided by \link{batch.summary}
}
\examples{
save.summary(b.150828)


}

