% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Screen-spots.R
\name{screen.spot.clusters}
\alias{screen.spot.clusters}
\title{Convert a list of screen spot coordinates into an image array and summary table}
\usage{
screen.spot.clusters(ss.coords, im.dims = c(1996, 1996), ignore.edges = 10)
}
\arguments{
\item{ss.coords}{Matrix of XY coordinates of identified screen spots, as returned by \code{\link{screen.spots}}}

\item{im.dims}{Dimensions of image on which to display coordinates}

\item{ignore.edges}{Integer: width of border around panel within which pixels are assumed to be dim because of their proximity to the edge, rather than because of a spot on the screen. Spots whose midpoint falls within this border will be removed. Default is 10.}
}
\value{
List containing an image array of the specified dimensions, and a data frame summarising each spot's midpoint and size.
}
\description{
Convert a list of screen spot coordinates into an image array and summary table
}
\examples{
image(screen.spot.clusters(screen.spots(pw.m[,,"white", "160430"]))$spots)

}

