% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_All-the-functions.R
\name{screen.spots.old}
\alias{screen.spots.old}
\title{Identify pixels covered by spots on screen}
\usage{
screen.spots.old(im, min.diam = 5, smooth.span = 1/15, midline = 1024.5,
  edge.crop = 10, coords = T)
}
\arguments{
\item{im}{Image array containing an image with x-ray exposure, in which non-responsive pixels are to be found.}

\item{min.diam}{Integer: minimum diameter of a dim spot, in pixels (used to set size of circular structuring element for morphological closing). Default is 5.}

\item{smooth.span}{Smoothing span to be used for Lowess. Default is 1/15 (this is a lower proportion than the standard default, since the aim is to 'skim over' any deviations from a smooth curve, not to fit them)}

\item{midline}{Numeric: y-coordinate of horizontal panel midline, dividing electrically separates panels. Default is 1024.5.}
}
\value{
Matrix of XY coordinates of pixels identified as probable screen spots
}
\description{
Find dim spots of a given size or larger in a single bright image, and return their x and y coordinates
}
\examples{
bp <- data.frame(screen.spots(pw.m[,,"white", "160430"]), type = "screen.spot")

}

