% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Screen-spots.R
\name{screen.spots}
\alias{screen.spots}
\title{Get screen spots in linear/white images}
\usage{
screen.spots(im, smooth.span = 1/15, min.diam = 5, midline = 1024.5,
  match.width = 100)
}
\description{
Screen spot identification in linear/white images
}

