% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Screen-spots.R
\name{screen.spots.xy}
\alias{screen.spots.xy}
\title{Get coordinates of pixels covered by spots on screen}
\usage{
screen.spots.xy(dt, smooth.span = 1/5, min.diam = 5, edge.width = 10,
  auto.threshold = T)
}
\arguments{
\item{dt}{Integer/string specifying image date to be imported}

\item{smooth.span}{Smoothing span to be used for Lowess. Default is 1/5}

\item{min.diam}{Integer: minimum diameter of a dim spot, in pixels (used to set size of circular structuring element for morphological closing). Default is 5.}

\item{edge.width}{Integer: dim spots whose midpoint falls within this distance of the detector edge will be removed from the list of screen defects.}
}
\value{
Data frame containing XY coordinates and type definition as 'screen spot'
}
\description{
Find dim spots of a given size or larger in a single white image, and return their x and y coordinates
}
\details{
Requires that \code{pw.m} be available in the global environment, as by running \link{\code{load.pixel.means()}}
}
\examples{
dt <- "160314"
bp <- rbind(get.dim.bright.px(res[, , "grey", dt]),
            screen.spots.xy(dt))

}

