% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{sd.colours}
\alias{sd.colours}
\title{Colour scheme for contour plot}
\usage{
sd.colours()
}
\value{
Vector of colours. When used in conjunction with  \link{\code{th.levels()}}, shades thresholded values.
}
\description{
Support function: returns a vector of colours to be used in contour plotting.
}
\examples{
    filled.contour(x = c(1:dim(data)[1]), y = c(1:dim(data)[2]),
    data,
    levels = sd.levels(mean(data), sd(data)),
    col = sd.colours())


}

