% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_All-the-functions.R
\name{sd.levels}
\alias{sd.levels}
\title{Set contour plot levels}
\usage{
sd.levels(data, midpoint = "mode")
}
\arguments{
\item{data}{Matrix of data to plot}

\item{midpoint}{String: use mean or median as midpoint? Default is "mean".}
}
\value{
Vector of calculated levels
}
\description{
Support function: returns a vector of levels to be used in contour plotting, cutting data at mean and +- 0.5, 1, 2, 3 sd.
}
\examples{
    filled.contour(x = c(1:dim(data)[1]), y = c(1:dim(data)[2]),
                   data,
                   levels = sd.levels(data))


}

