% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{shading.corrected}
\alias{shading.corrected}
\title{Apply shading correction}
\usage{
shading.corrected(im)
}
\arguments{
\item{im}{Set of pixelwise means for black, grey and white acquisitions}
}
\description{
Calculate a shading-corrected image from a set of blanks
}

