% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{show.panels}
\alias{show.panels}
\title{Add panel edges to pixel image}
\usage{
show.panels(left.crop = 2, lower.crop = 32, width = 128, height = 1024)
}
\arguments{
\item{left.crop}{Number of pixels cropped from left-hand edge of image. Default is 2.}

\item{width}{Width of each panel, in pixels. Default is 128.}

\item{height}{Height of each panel, in pixels. Default is 1024.}

\item{left.crop}{Number of pixels cropped from lower edge of image. Default is 32.}
}
\description{
Add lines showing the edges of individual detector panels to a pixel value image (eg. created using \link{pixel.image})
}
\details{
Each detector panel outputs 128 x 1024 pixels, with some cropping at each edge of the combined image.
}
\examples{
pixel.image(pw.mean)
show.panels()


}

