% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_All-the-functions.R
\name{smooth.lines}
\alias{smooth.lines}
\title{Smooth lines}
\usage{
smooth.lines(im, sm.size = 11, horizontal = F, min.segment = 6)
}
\arguments{
\item{im}{2d image array to be convolved.}

\item{sm.size}{Length of (linear) kernel to use. Default is 11.}

\item{horizontal}{Boolean: look for horizontal lines (T) instead of vertical (F)? Default is F, look for vertical lines.}

\item{min.segment}{Integer: minimum number of pixels to accept as a line after smoothing.}
}
\value{
Smoothed and thresholded image
}
\description{
Use a focal window to smooth line segments and remove short fragments.
}

