% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-fitting.R
\name{spot.lm}
\alias{spot.lm}
\title{Linear regression over distance from centre}
\usage{
spot.lm(image, order = 2, centre = c(1023.5, 992.5), robust = F)
}
\arguments{
\item{image}{Single-layer (1996x1996) array image to be fitted.}

\item{order}{Order of polynomial to include. Default is 2}

\item{centre}{Vector of coordinates of centre of spot. Default is centre of uncropped detector, c(1023.5, 992.5)}

\item{robust}{Boolean: use robust model fitting (\link{\code{rlm}})? Default is F (use \link{\code{lm}} for regression).}
}
\value{
Linear model fitted to given image
}
\description{
Fit a circular spot model with linear gradients to the image, including polynomial terms of z as explanatory variables
}
\examples{
circ.lm <- spot.lm(pw.m)
circ.res <- matrix(circ.lm$residuals, ncol = 1996)
pixel.image(circ.res)

}

