% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_All-the-functions.R
\name{subpanels}
\alias{subpanels}
\title{Split image array into subpanels}
\usage{
subpanels(data, left.pad = 2, upper.pad = 20)
}
\arguments{
\item{data}{Single-layer 1996x1996 image array}

\item{left.pad}{Integer numer of pixels cropped from left-hand edge of image. Default is 2}

\item{upper.pad}{Integer numer of pixels cropped from upper edge of image. Default is 20}
}
\description{
Take a single image layer and return 32-layer labelled subpanel array to allow easier computation over panels.
}
\examples{
load.images(150828, "black")
pw.m <- pixelwise.mean(b.150828)

zz <- subpanels(pw.m)
pixel.image(zz[,,"U8"], break.levels = sd.levels(b.150828))

}

