% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_All-the-functions.R
\name{summarise.images}
\alias{summarise.images}
\title{Calculate summary statistics for all images}
\usage{
summarise.images(fpath = "/home/clair/Documents/Pixels/Image-data/")
}
\arguments{
\item{fpath}{Path to top level of image directory. Default is "/home/clair/Documents/Pixels/Image-data/"}
}
\value{
Data frame containing summaries of pixel values: mean and standard deviation, max, min, quartiles and IQR.
}
\description{
Searches through folder structure, extracts all .tif images, and stores the summary statistics for each in a data frame.
}
\examples{
df <- summarise.images()


}

