% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Line-detection.R
\name{summarise.lines}
\alias{summarise.lines}
\title{Summarise dimensions of line segments identified}
\usage{
summarise.lines(im, rows = F, filter = F, filter.at = list(cover = 0.5,
  filled = 20))
}
\arguments{
\item{im}{2d image array after convolution, thresholding and smoothing}

\item{rows}{Boolean: summarise over rows (T) or columns (F)? Default is columns.}

\item{filter}{Boolean: return the full data frame (F) or only longer lines (T)? Default is full data frame.}

\item{filter.at}{List of numeric filtering parameters: 'cover' is minimum proportion of line that must be filled to qualify as a long line (default: 0.5), 'filled' is minimum length (default: 20).}
}
\value{
Data frame containing maximum and minimum row identified as part of a line segment, and the range and proportion of each column covered by line segments.
}
\description{
After identifying potential bright/dim lines/rows using \link{\code{find.lines}}, produce a table of summary information about each column found.
}

