% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_All-the-functions.R
\name{summarise.profiles}
\alias{summarise.profiles}
\title{Summarise image profile data}
\usage{
summarise.profiles(fpath = "/home/clair/Documents/Pixels/Image-data/")
}
\arguments{
\item{fpath}{Path to top level of image directory. Default is "/home/clair/Documents/Pixels/Image-data/"}
}
\value{
Data frame containing summaries of image profile data.
}
\description{
Searches through folder structure, extracts all .xml profiles, and summarises into a data frame.
}
\details{
If NA appears in data frame, the acquisitions for that batch do not all share the same value of that attribute, and should be investigated individually.
}
\examples{
df <- summarise.profiles()


}

