% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_All-the-functions.R
\name{superclusters}
\alias{superclusters}
\title{Identify superclusters of pixels}
\usage{
superclusters(cl, excl = c(""))
}
\arguments{
\item{cl}{Bad pixel map or clustered bad pixel map: a matrix or data frame containing coordinates of bad pixels and column "type" defining bad pixel type, with cluster ID and count if already clustered by \link{\code{cluster.by.type}}}

\item{Vector}{of bad pixel types to exclude from superclusters. Default is to exclude screen spots and edge pixels.}
}
\value{
Data frame containing bad pixel coordinates and type, along with ID and size of clusters and superclusters to which each is assigned, and a shape classification based on these.
}
\description{
Identify superclusters of bad pixels: small clusters of adjacent bad pixels of different types.
}
\details{
Only clusters of size 9 or less will be considered as part of a supercluster, with the exception of bright or dim columns or rows identified. To exclude bright or dim lines, add those types to the 'excl' parameter.
}
\examples{
# should both give same result
zz <- superclusters(cluster.by.type(bp))
zz <- superclusters(bp)

}

