% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import.R
\name{system.map}
\alias{system.map}
\title{Import system-generated bad pixel map}
\usage{
system.map(dt, fpath = "./Image-data/",
  img = paste0("./02_Objects/images/pwm-", dt, ".rds"))
}
\arguments{
\item{dt}{Date of image for which pixel map is to be imported.}

\item{fpath}{Path to image folder}
}
\value{
Data.frame containing x and y coordinates
}
\description{
Extract coordinates of bad pixels from system-generated bad pixel map
}

