% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thresholding.R
\name{threshold.pixels}
\alias{threshold.pixels}
\title{Load one day's acquisitions from one channel}
\usage{
threshold.pixels(data, too.bright = 1.5, too.dim = 0.45, noisy = 6)
}
\arguments{
\item{data}{Array containing a sequence of images from a single channel}

\item{too.bright}{Numeric value to use as threshold for underperforming bright pixels: the multiple of median value above which a pixel is deemed to be unacceptably bright. Default is 1.5 x median.}

\item{too.dim}{Numeric value to use as threshold for underperforming dark pixels: the multiple of median value below which a pixel is deemed to be unacceptably bright. Default is 0.45 x median.}

\item{noisy}{Numeric value to use as threshold for noisy pixels: the number of median pixel SDs above which a pixel is deemed to be noisy. Default is 6.}
}
\value{
Data frame containing coordinates and classifications of all underperforming pixels in the given array.
}
\description{
Import one day's images from a single channel into an array. If the function is called by the user, the data is loaded into an object named automatically by the function. If called by another function, will return an array.
}
\examples{
bp.b.150828 <- threshold.pixels(b.150828)


}

