% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BP-parametric.R
\name{bp.parametric.quantiles}
\alias{bp.parametric.quantiles}
\title{Identify 'bad' pixels using quantiles of residuals after parametric fitting}
\usage{
bp.parametric.quantiles(img.date, batch, circ.o = 2, panel.o = 1,
  lq = 0.001, uq = 0.999, dist = 1.5, details = F)
}
\arguments{
\item{img.date}{Integer/string specifying image date to be imported}

\item{batch}{String specifying colour batch to be imported (black, grey or white)}

\item{circ.o}{Integer (or NA if no spot model is required) specifying order of polynomial model to fit to circular spot.}

\item{panel.o}{Integer (or NA if no per-panel model is required) specifying order of polynomial model to fit to panel X and Y coordinates.}

\item{lq}{Lower quartile at which to start looking for 'outliers'. Default is 0.001}

\item{uq}{Upper quartile at which to start looking for 'outliers'. Default is 0.999}

\item{dist}{Multiples of IQR away from lq/uq at which a point is classified as 'bad'.}

\item{details}{Boolean: return lists of identified 'bad pixels' (T) or only summary (F)? Default is F.}
}
\description{
Load a specified image and fit parametric circular spot and panelwise gradients, then cut residuals at specified quantiles to classify 'bad' pixels
}
\details{
Requires that pixelwise mean files are already loaded, as by \link{\code{load.pixel.means}}
}
\examples{
load.pixel.means()
zz <- rbind(bp.parametric.quantiles(160314, "white"),
            bp.lowess.quantiles(160314, "white"))

}

