% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_All-the-functions.R
\name{cluster.by.type}
\alias{cluster.by.type}
\title{Identify clusters of pixels}
\usage{
cluster.by.type(bpm)
}
\arguments{
\item{bpm}{Bad pixel map: a matrix or data frame containing coordinates of bad pixels and column "type" defining bad pixel type ("hot", "bright", "dim", "dead").}
}
\value{
Data frame containing bad pixel coordinates and type, along with ID and size of cluster to which each is assigned.
}
\description{
Identify clusters of adjacent pixels of each type in a bad pixel map.
}
\examples{
zz <- cluster.by.type(bpm)

}

