% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import.R
\name{load.daily}
\alias{load.daily}
\title{Load one day's acquisitions from all batches}
\usage{
load.daily(img.date, fpath = "/home/clair/Documents/Pixels/Image-data/")
}
\arguments{
\item{img.date}{Date of images to import, in format yymmdd.}

\item{fpath}{Path to top level of stored images. Default is "/home/clair/Documents/Pixels/Image-data/"}
}
\description{
Import one day's images from into  1996x1996x20x3 array.
}
\examples{
img.150828 <- load.daily(150828)


}

