% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import.R
\name{load.pixel.means}
\alias{load.pixel.means}
\title{Create array of all pixelwise means}
\usage{
load.pixel.means(fpath = "./Other-data/")
}
\arguments{
\item{fpath}{Path to top level of stored images. Default is "./Other-data/"}
}
\description{
Load all pixelwise mean summaries into a single labelled array.
}
\examples{
load.pixel.means()

}

