% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_All-the-functions.R
\name{load.pixel.means}
\alias{load.pixel.means}
\title{Create array of all pixelwise means}
\usage{
load.pixel.means(fpath = "./02_Objects/images", acq.list)
}
\arguments{
\item{fpath}{Path to top level of stored .rds objects. Default is "./02_Objects/images"}

\item{acq.list}{Vector containing acquisitions to import. If left blank, will import every file in the target folder named "pwm-xxxxxx.rds".}
}
\description{
Load all pixelwise mean summaries (already created and stored as .rds objects) into a single labelled array.
}
\details{
RDS objects for import can be downloaded from https://github.com/ClairBee/Pixels/tree/master/02_Objects/images
}

