% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_All-the-functions.R
\name{load.profiles}
\alias{load.profiles}
\title{Load xml profiles of one day's acquisitions from one batch}
\usage{
load.profiles(img.date, batch,
  fpath = "/home/clair/Documents/Pixels/Image-data/")
}
\arguments{
\item{img.date}{Date of images to import, in format yymmdd.}

\item{batch}{Specify batch to import: black, grey or white.}

\item{fpath}{Path to top level of stored images. Default is "/home/clair/Documents/Pixels/Image-data/"}
}
\value{
Data frame containing all xml profile data for the specified image set.
}
\description{
Support function: import xml profiles of one day's images into a single data frame
}
\examples{
b.150828.xml <- load.profiles(150828, "black")


}

