% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{no.response}
\alias{no.response}
\title{Identify pixels with no x-ray response}
\usage{
no.response(dt, limit = 0.01)
}
\arguments{
\item{dt}{String or integer date, format yymmdd.}

\item{limit}{Quantile of black pixel population to be used as cutoff.}
}
\value{
Matrix of coordinates of non-responsive pixels
}
\description{
Identify pixels whose behaviour in the grey or white images is the same as in the black images (no response to source)
}
\details{
Requires existence of \code{pw.m} in global environment.
}
\examples{
qq <- no.response(160314, limit = 0.001)
}

