% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_All-the-functions.R
\name{panel.lm}
\alias{panel.lm}
\title{Per-panel linear regression}
\usage{
panel.lm(image, terms = "x + y", robust = F, left.pad = 2,
  upper.pad = 20, rotate.lower = F)
}
\arguments{
\item{image}{Single-layer array image (1996x1996) to be fitted.}

\item{terms}{String specifying terms to be fitted to each subpanel. Default is "x + y", fitting a linear model to the x and y coordinates of each subpanel without considering interactions.}

\item{robust}{Boolean: use robust regression (\link{\code{rlm}})? Default is F (use \link{\code{lm}} for regression).}

\item{left.pad}{Integer value: how many pixels are cropped from left image edge? Default is 2.}

\item{upper.pad}{Integer value: how many pixels are cropped from upper image edge? Default is 20.}

\item{rotate.lower}{boolean: fit a model to the lower panels as they appear in the image, or rotate by 180 degrees to align amplifier coordinates with those of top panels? Default is F (fit models to image as is)}
}
\value{
List containing the terms applied, a matrix of fitted values, and a matrix of model coefficients for each panel.
}
\description{
Fit a linear gradient over each of the 32 subpanels within the image
}
\examples{
panel.lm <- panel.lm(circ.res, "poly(x, 2) + y")    # finds coefficients of intercept, x, x^2, y.
panel.lm <- panel.lm(circ.res, "y")                 # finds coefficients of intercept and y.
panel.res <- circ.res - panel.lm$fitted.values
pixel.image(panel.res)
}

