% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{pixel.image}
\alias{pixel.image}
\title{Display pixel values as image}
\usage{
pixel.image(data, title = "", x.range = c(1:nrow(data)),
  y.range = c(1:ncol(data)), break.levels = th.levels(data), panels = F,
  x.lab = "", y.lab = "", ...)
}
\arguments{
\item{data}{2-dimensional matrix containing values to be plotted}

\item{title}{String containing title to be printed with plot}

\item{x.range}{Vector range showing x-range of cells to be included in plot}

\item{y.range}{Vector range showing y-range of cells to be included in plot}

\item{break.levels}{Vector of values used as breakpoints when binning values}

\item{midpoint}{Either 'mean' or 'median'. Default is 'mean'}
}
\description{
Create an image with pixels shaded according to their distance from the mean value. Displays at lower resolution than contour plot (plots are 20% of the size), but much quicker (10% of time taken).
}
\details{
Pixels more than 2sd above the mean are red, pink or purple. Pixels within 2sd of the mean are orange, yellow or green. Pixels more than 2sd below the mean are blue.
}
\examples{
pixel.image(pw.mean)


}

