% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{pixel.plot}
\alias{pixel.plot}
\title{Plot given coordinates}
\usage{
pixel.plot(px, xlim = c(0, 2048), ylim = c(0, 2048), pch = 15,
  panels = F, cex = 0.4, main = "", xlab = "", ylab = "", ...)
}
\arguments{
\item{px}{Coordinates of pixels to be plotted}
}
\description{
Plot pixel map
}
\examples{
pixel.plot(which(pw.m[,,"white"] < 15000, arr.ind = T))

}

