% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{pixel.plot}
\alias{pixel.plot}
\title{Plot defective pixels}
\usage{
pixel.plot(px, xlim = c(0, 2048), ylim = c(0, 2048), pch = 15,
  panels = F, main = "", xlab = "", ylab = "", ...)
}
\arguments{
\item{px}{Coordinates of pixels to be plotted}
}
\description{
Plot bad pixel map on 2048 x 2048 grid
}
\examples{
pixel.plot(which(pw.mean > 25000, arr.ind = T))


}

