% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{plot.bad.px}
\alias{plot.bad.px}
\title{Plot bad pixels}
\usage{
plot.bad.px(bpm, pch = 15, xlab = "", ylab = "", ...)
}
\arguments{
\item{bpm}{Data frame containing xy coordinates of pixels to be plotted, and a field "type" specifying the category of each bad pixel identified.}
}
\description{
Plot the coordinates and types of a bad pixel map
}

