% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conversion.R
\name{r2m}
\alias{r2m}
\title{Convert raster to matrix}
\usage{
r2m(rast)
}
\arguments{
\item{rast}{Single-layer RasterLayer object to be converted to a 2d array}
}
\value{
Single-layer image array
}
\description{
Convert a raster to an image array, maintaining the correct orientation for plotting
}

