% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Line-detection.R
\name{summarise.lines}
\alias{summarise.lines}
\title{Summarise dimensions of line segments identified}
\usage{
summarise.lines(im, horizontal = F, midline = 1024.5)
}
\arguments{
\item{im}{2d image array after convolution, thresholding and smoothing}

\item{horizontal}{Boolean: summarise over rows (T) or columns (F)? Default is columns.}
}
\value{
Data frame containing maximum and minimum row identified as part of a line segment, and the range and proportion of each column covered by line segments.
}
\description{
After identifying potential bright/dim lines/rows using \link{\code{find.lines}}, produce a table of summary information about each column found.
}

