% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{th.levels}
\alias{th.levels}
\title{Set contour plot levels}
\usage{
th.levels(data)
}
\arguments{
\item{data}{Matrix of data to plot}
}
\value{
Vector of calculated levels
}
\description{
Support function: returns a vector of levels to be used in contour plotting, cutting data at mean and +- 1,2,3,4.5,6 mad.
}
\examples{
    filled.contour(x = c(1:dim(data)[1]), y = c(1:dim(data)[2]),
                   data,
                   levels = sd.levels(data))

}

